Refactor DistributionEventHandler into v2 package
