+++ pilot/pkg/model/context.go
@@ -71,8 +71,6 @@ type Environment struct {
 	// CONFIG AND PUSH
 	PushContext *PushContext
 
-	StatusReporter DistributionEventHandler
-
 	// DomainSuffix provides a default domain for the Istio server.
 	DomainSuffix string
 }
