+++ pilot/pkg/proxy/envoy/v2/ads.go
@@ -234,8 +234,8 @@ func (s *DiscoveryServer) StreamAggregatedResources(stream ads.AggregatedDiscove
 				}
 				defer s.removeCon(con.ConID)
 			}
-			if s.Env.StatusReporter != nil {
-				s.Env.StatusReporter.RegisterEvent(con.ConID, discReq.TypeUrl, discReq.ResponseNonce)
+			if s.StatusReporter != nil {
+				s.StatusReporter.RegisterEvent(con.ConID, discReq.TypeUrl, discReq.ResponseNonce)
 			}
 
 			// Based on node metadata a different generator was selected, use it instead of the default
@@ -605,11 +605,11 @@ func (s *DiscoveryServer) pushConnection(con *XdsConnection, pushEv *XdsEvent) e
 			adsLog.Debugf("Skipping push to %v, no updates required", con.ConID)
 		}
 
-		if s.Env.StatusReporter != nil {
+		if s.StatusReporter != nil {
 			// inform distribution status reporter that this connection has been updated, because it effectively has
 			go func() {
 				for _, typeURL := range []string{ClusterType, ListenerType, RouteType, EndpointType} {
-					s.Env.StatusReporter.RegisterEvent(con.ConID, typeURL, pushEv.noncePrefix)
+					s.StatusReporter.RegisterEvent(con.ConID, typeURL, pushEv.noncePrefix)
 				}
 			}()
 		}
@@ -642,8 +642,8 @@ func (s *DiscoveryServer) pushConnection(con *XdsConnection, pushEv *XdsEvent) e
 		if err != nil {
 			return err
 		}
-	} else if s.Env.StatusReporter != nil {
-		s.Env.StatusReporter.RegisterEvent(con.ConID, ClusterType, pushEv.noncePrefix)
+	} else if s.StatusReporter != nil {
+		s.StatusReporter.RegisterEvent(con.ConID, ClusterType, pushEv.noncePrefix)
 	}
 
 	if len(con.Clusters) > 0 && pushTypes[EDS] {
@@ -651,24 +651,24 @@ func (s *DiscoveryServer) pushConnection(con *XdsConnection, pushEv *XdsEvent) e
 		if err != nil {
 			return err
 		}
-	} else if s.Env.StatusReporter != nil {
-		s.Env.StatusReporter.RegisterEvent(con.ConID, EndpointType, pushEv.noncePrefix)
+	} else if s.StatusReporter != nil {
+		s.StatusReporter.RegisterEvent(con.ConID, EndpointType, pushEv.noncePrefix)
 	}
 	if con.LDSWatch && pushTypes[LDS] {
 		err := s.pushLds(con, pushEv.push, currentVersion)
 		if err != nil {
 			return err
 		}
-	} else if s.Env.StatusReporter != nil {
-		s.Env.StatusReporter.RegisterEvent(con.ConID, ListenerType, pushEv.noncePrefix)
+	} else if s.StatusReporter != nil {
+		s.StatusReporter.RegisterEvent(con.ConID, ListenerType, pushEv.noncePrefix)
 	}
 	if len(con.Routes) > 0 && pushTypes[RDS] {
 		err := s.pushRoute(con, pushEv.push, currentVersion)
 		if err != nil {
 			return err
 		}
-	} else if s.Env.StatusReporter != nil {
-		s.Env.StatusReporter.RegisterEvent(con.ConID, RouteType, pushEv.noncePrefix)
+	} else if s.StatusReporter != nil {
+		s.StatusReporter.RegisterEvent(con.ConID, RouteType, pushEv.noncePrefix)
 	}
 	proxiesConvergeDelay.Record(time.Since(pushEv.start).Seconds())
 	return nil
@@ -786,8 +786,8 @@ func (s *DiscoveryServer) removeCon(conID string) {
 	}
 
 	xdsClients.Record(float64(len(s.adsClients)))
-	if s.Env.StatusReporter != nil {
-		go s.Env.StatusReporter.RegisterDisconnect(conID, []string{ClusterType, ListenerType, RouteType, EndpointType})
+	if s.StatusReporter != nil {
+		go s.StatusReporter.RegisterDisconnect(conID, []string{ClusterType, ListenerType, RouteType, EndpointType})
 	}
 }
